Sistema de Asistencia -Minor fix en el menu -Correciones en mesas y modulos para evitar borrarlos cuando poseen datos -Agregado del sistema de asistencia
Pendiente la carga de inscriptos y asignacion a las mesas

+++ admin/rrhh_eventos_mesas.php
@@ -16,8 +16,10 @@ if (!$_SESSION['logged'] || !$esadmin) {
 
 if (isset($_POST['evento'])){
 	$evento=intval($_POST['evento']);
-} else {
-	$evento=0;
+} else if (isset($_GET['evento'])) {
+		$evento=intval($_GET['evento']);
+	} else {
+		$evento=0;
 }
 
 if (isset($_POST['mesa'])){
@@ -54,18 +56,26 @@ if (isset($_POST['mesa_nombre']) && isset($_POST['coordinador_1']) && isset($_PO
 	$m_id = mysql_real_escape_string(substr(htmlspecialchars($_POST['mesa']),0,10));
 	$m_ev = mysql_real_escape_string(substr(htmlspecialchars($_POST['evento']),0,10));
 
-	$sql = "DELETE FROM rtc_eventos_mesa WHERE id='$m_id' AND evento_id='$m_ev'";
+	$sql = "SELECT * FROM rtc_eventos_mesa_modulos WHERE mesa_id='$m_id' LIMIT 1";
 	$result = mysql_query($sql);
-	$mesa=0;
+	$row = mysql_fetch_assoc($result);
+	if ($row) {
+		echo "<div class=\"muestra_alarma\">Imposible borrar porque la mesa tiene modulos ingresados</div>";
+	} else {
+		$sql = "DELETE FROM rtc_eventos_mesa WHERE id='$m_id' AND evento_id='$m_ev'";
+		$result = mysql_query($sql);
+		$mesa=0;
+	}
 }
 
 ?>
+<div><h2>Edicion de Mesas</h2></div>
 <div>
 <form id="form1" name="form1" method="POST" action="rrhh_eventos_mesas.php">Seleccione un evento:
 <?php
 	$sql1 = "SELECT * FROM rtc_eventos ORDER BY nombre";
 	$resultado = mysql_query($sql1);
-	echo "<select name=\"evento\" id=\"evento\">";
+	echo "<select name=\"evento\" id=\"evento\" onchange=\"location.href='rrhh_eventos_mesas.php?evento='+this.value\" >";
 	echo "<option value=\"0\" selected > </option>";
 	while ($rowtmp = mysql_fetch_assoc($resultado))
 	{
@@ -73,7 +83,6 @@ if (isset($_POST['mesa_nombre']) && isset($_POST['coordinador_1']) && isset($_PO
 	}
 	echo "</select>";
 ?>
-<input type="submit" name="button" id="button" value="Enviar" />
 </form>
 </div>
 <?php if ($evento!='0') {?>
